# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/projects/{projectId}/prompts/{promptId}/assets/{assetKey}": {
      "get": {
        "operationId": "PromptAssetController_findOne",
        "parameters": [
          {
            "name": "promptId",
            "required": true,
            "in": "path",
            "description": "ID (slug) del prompt padre",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID del proyecto",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assetKey",
            "required": true,
            "in": "path",
            "description": "Key única del asset dentro del prompt",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Asset encontrado con detalles."
          },
          "404": {
            "description": "Proyecto o Asset no encontrado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtiene un prompt asset por su key dentro de un prompt específico",
        "tags": [
          "Prompt Assets (for a specific Prompt)"
        ]
      },
      "patch": {
        "operationId": "PromptAssetController_update",
        "parameters": [
          {
            "name": "promptId",
            "required": true,
            "in": "path",
            "description": "ID (slug) del prompt padre",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID del proyecto",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assetKey",
            "required": true,
            "in": "path",
            "description": "Key única del asset a actualizar",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePromptAssetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Asset actualizado."
          },
          "400": {
            "description": "Datos inválidos."
          },
          "404": {
            "description": "Proyecto o Asset no encontrado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualiza metadatos de un prompt asset (nombre, descripción, etc.) dentro de un prompt",
        "tags": [
          "Prompt Assets (for a specific Prompt)"
        ]
      },
      "delete": {
        "operationId": "PromptAssetController_remove",
        "parameters": [
          {
            "name": "promptId",
            "required": true,
            "in": "path",
            "description": "ID (slug) del prompt padre",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID del proyecto",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assetKey",
            "required": true,
            "in": "path",
            "description": "Key única del asset a eliminar",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Asset eliminado."
          },
          "404": {
            "description": "Proyecto o Asset no encontrado."
          },
          "409": {
            "description": "Conflicto al eliminar (revisar relaciones sin Cascade)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Elimina un prompt asset (y sus versiones/traducciones por Cascade) dentro de un prompt",
        "tags": [
          "Prompt Assets (for a specific Prompt)"
        ]
      }
    },
...
    "/api/projects/{projectId}/prompts/{promptId}/versions/{versionTag}/translations/{languageCode}": {
      "get": {
        "operationId": "PromptTranslationController_findOneByLanguage",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "promptId",
            "required": true,
            "in": "path",
            "description": "Prompt CUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "versionTag",
            "required": true,
            "in": "path",
            "description": "Version Tag (e.g., 1.0.0) or \"latest\" to get the most recent version",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "languageCode",
            "required": true,
            "in": "path",
            "description": "Language code (e.g., es-ES)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resolveAssets",
            "required": false,
            "in": "query",
            "description": "Whether to resolve asset placeholders. Defaults to false.",
            "schema": {
              "example": "true",
              "type": "boolean"
            }
          },
          {
            "name": "environmentId",
            "required": false,
            "in": "query",
            "description": "Environment ID for context.",
            "schema": {
              "example": "dev-env-123",
              "type": "string"
            }
          },
          {
            "name": "regionCode",
            "required": false,
            "in": "query",
            "description": "Region code for context (e.g., for asset translations - overrides languageCode for assets if provided).",
            "schema": {
              "example": "us-east-1",
              "type": "string"
            }
          },
          {
            "name": "variables",
            "required": false,
            "in": "query",
            "description": "JSON stringified object of variables for substitution.",
            "schema": {
              "example": "{\"varName\": \"value\"}",
              "type": "string"
            }
          },
          {
            "name": "processed",
            "required": false,
            "in": "query",
            "description": "Whether to return the processed prompt with all references and variables resolved. Defaults to false.",
            "schema": {
              "example": "true",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Translation found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePromptTranslationDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden Access to Project."
          },
          "404": {
            "description": "Project, Prompt, Version, or Translation not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a specific translation by language code for a prompt version. Allows resolving assets.",
        "tags": [
          "Prompt Translations (within Project/Prompt/Version)"
        ]
      },
      "patch": {
        "operationId": "PromptTranslationController_update",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "promptId",
            "required": true,
            "in": "path",
            "description": "Prompt CUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "versionTag",
            "required": true,
            "in": "path",
            "description": "Version Tag (e.g., 1.0.0) or \"latest\" to get the most recent version",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "languageCode",
            "required": true,
            "in": "path",
            "description": "Language code of the translation to update",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePromptTranslationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Translation updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePromptTranslationDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden Access to Project."
          },
          "404": {
            "description": "Project, Prompt, Version, or Translation not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a specific translation by language code for a prompt version",
        "tags": [
          "Prompt Translations (within Project/Prompt/Version)"
        ]
      },
      "delete": {
        "operationId": "PromptTranslationController_remove",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "promptId",
            "required": true,
            "in": "path",
            "description": "Prompt CUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "versionTag",
            "required": true,
            "in": "path",
            "description": "Version Tag (e.g., 1.0.0) or \"latest\" to get the most recent version",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "languageCode",
            "required": true,
            "in": "path",
            "description": "Language code of the translation to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Translation deleted."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden Access to Project."
          },
          "404": {
            "description": "Project, Prompt, Version, or Translation not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a specific translation by language code for a prompt version",
        "tags": [
          "Prompt Translations (within Project/Prompt/Version)"
        ]
      }
    },
...
    "/api/projects/{projectId}/prompts/{promptId}/assets/{assetKey}/versions/{versionTag}/translations": {
      "post": {
        "operationId": "AssetTranslationController_create",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID of the Project the Prompt belongs to",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "promptId",
            "required": true,
            "in": "path",
            "description": "ID (slug) of the Prompt",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assetKey",
            "required": true,
            "in": "path",
            "description": "Key of the PromptAsset",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "versionTag",
            "required": true,
            "in": "path",
            "description": "Tag of the PromptAssetVersion",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAssetTranslationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Translation created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAssetTranslationDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden Access to Project."
          },
          "404": {
            "description": "Project, Asset, or Version not found."
          },
          "409": {
            "description": "Translation for this language already exists for this version."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a translation for a specific asset version",
        "tags": [
          "Asset Translations (Project > Prompt > Asset > Version > Translation)"
        ]
      },
      "get": {
        "operationId": "AssetTranslationController_findAll",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID of the Project the Prompt belongs to",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "promptId",
            "required": true,
            "in": "path",
            "description": "ID (slug) of the Prompt",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assetKey",
            "required": true,
            "in": "path",
            "description": "Key of the PromptAsset",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "versionTag",
            "required": true,
            "in": "path",
            "description": "Tag of the PromptAssetVersion",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of translations.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CreateAssetTranslationDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden Access to Project."
          },
          "404": {
            "description": "Project, Asset, or Version not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all translations for a specific asset version",
        "tags": [
          "Asset Translations (Project > Prompt > Asset > Version > Translation)"
        ]
      }
    },
...
  },
  "info": {
    "title": "japm.app API",
    "description": "API for the japm.app Prompt Engineering application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "InitialTranslationDto": {
        "type": "object",
        "properties": {
          "languageCode": {
            "type": "string",
            "minLength": 2,
            "maxLength": 10,
            "description": "ISO language code (e.g., es, en)"
          },
          "promptText": {
            "type": "string",
            "description": "Translated prompt text"
          }
        },
        "required": [
          "languageCode",
          "promptText"
        ]
      },
...
      "UpdatePromptAssetDto": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string",
            "description": "Categoría para organizar assets (e.g., Saludos, Despedidas)",
            "example": "Saludos"
          },
          "tenantId": {
            "type": "string",
            "description": "ID del tenant al que pertenece este asset",
            "example": "tenant-cuid-xxxx"
          },
          "initialTranslations": {
            "description": "Traducciones iniciales para diferentes idiomas",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InitialTranslationDto"
            }
          },
          "enabled": {
            "type": "boolean",
            "description": "Activa o desactiva el asset"
          }
        }
      },
...
      "CreatePromptTranslationDto": {
        "type": "object",
        "properties": {
          "languageCode": {
            "type": "string",
            "minLength": 5,
            "maxLength": 5,
            "description": "Código de idioma para esta traducción (formato xx-XX)",
            "example": "es-ES"
          },
          "promptText": {
            "type": "string",
            "description": "Texto del prompt traducido a este idioma"
          }
        },
        "required": [
          "languageCode",
          "promptText"
        ]
      },
      "UpdatePromptTranslationDto": {
        "type": "object",
        "properties": {
          "promptText": {
            "type": "string",
            "description": "Texto del prompt traducido a este idioma"
          }
        }
      },
...
      "CreateAssetTranslationDto": {
        "type": "object",
        "properties": {
          "languageCode": {
            "type": "string",
            "minLength": 5,
            "maxLength": 5,
            "pattern": "/^[a-z]{2}-[A-Z]{2}$/",
            "description": "Código de idioma para esta traducción (formato xx-XX)",
            "example": "es-ES"
          },
          "value": {
            "type": "string",
            "description": "Valor del asset traducido a este idioma"
          }
        },
        "required": [
          "languageCode",
          "value"
        ]
      },
...
    }
  }
}